split endpoints
